refactor: extract input validation
